+++ runtime/Java/src/org/antlr/v4/runtime/atn/LexerATNSimulator.java
@@ -67,7 +67,7 @@ public class LexerATNSimulator extends ATNSimulator {
 	 *  can simply return the predicted token type.
 	 */
 	protected static class SimState {
-		protected int index = -1;
+		protected long index = -1;
 		protected int line = 0;
 		protected int charPos = -1;
 		protected DFAState dfaState;
@@ -88,7 +88,7 @@ public class LexerATNSimulator extends ATNSimulator {
 	 *  DFA did not have a previous accept state. In this case, we use the
 	 *  ATN-generated exception object.
 	 */
-	protected int startIndex = -1;
+	protected long startIndex = -1;
 
 	/** line number 1..n within the input */
 	protected int line = 1;
@@ -131,7 +131,7 @@ public class LexerATNSimulator extends ATNSimulator {
 	public int match(@NotNull CharStream input, int mode) {
 		match_calls++;
 		this.mode = mode;
-		int mark = input.mark();
+		long mark = input.mark();
 		try {
 			this.startIndex = input.index();
 			this.prevAccept.reset();
@@ -356,7 +356,7 @@ public class LexerATNSimulator extends ATNSimulator {
 	}
 
 	protected void accept(@NotNull CharStream input, int ruleIndex, int actionIndex,
-						  int index, int line, int charPos)
+						  long index, int line, int charPos)
 	{
 		if ( debug ) {
 			System.out.format(Locale.getDefault(), "ACTION %s:%d\n", recog != null ? recog.getRuleNames()[ruleIndex] : ruleIndex, actionIndex);
@@ -555,8 +555,8 @@ public class LexerATNSimulator extends ATNSimulator {
 
 		int savedCharPositionInLine = charPositionInLine;
 		int savedLine = line;
-		int index = input.index();
-		int marker = input.mark();
+		long index = input.index();
+		long marker = input.mark();
 		try {
 			consume(input);
 			return recog.sempred(null, ruleIndex, predIndex);
